% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiforest.R
\name{calculate_multiforest_layout}
\alias{calculate_multiforest_layout}
\title{Calculate table layout for multiforest plots}
\usage{
calculate_multiforest_layout(
  to_show_exp_clean,
  show_n,
  show_events,
  indent_predictor,
  show_predictor = TRUE,
  effect_abbrev,
  font_size,
  table_width = 0.6,
  rangeb,
  center_padding,
  ci_pct = 95
)
}
\description{
Computes column widths and positions for the table portion of a multiforest plot.
}
\keyword{internal}
