% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifit.R
\name{combine_multifit_results}
\alias{combine_multifit_results}
\title{Combine multivariate results}
\usage{
combine_multifit_results(all_results, columns)
}
\arguments{
\item{all_results}{List of results from fit_one_outcome calls.}

\item{columns}{Character specifying which columns to include.}
}
\value{
Combined data.table.
}
\description{
Internal helper to combine unadjusted and/or adjusted results from 
multiple outcomes into a single data.table.
}
\keyword{internal}
