% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforest.R
\name{is_multifit_result}
\alias{is_multifit_result}
\title{Check if object is a multifit result}
\usage{
is_multifit_result(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical indicating if x is a multifit result.
}
\description{
Internal helper to detect multifit output objects.
}
\keyword{internal}
