% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utils.R
\name{is_surv_outcome}
\alias{is_surv_outcome}
\title{Check if outcome is a Surv() expression}
\usage{
is_surv_outcome(outcome)
}
\arguments{
\item{outcome}{Character string of the outcome specification}
}
\value{
Logical
}
\description{
Check if outcome is a Surv() expression
}
\keyword{internal}
