% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforest.R
\name{is_uniscreen_result}
\alias{is_uniscreen_result}
\title{Check if object is a uniscreen result}
\usage{
is_uniscreen_result(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical indicating if x is a uniscreen result.
}
\description{
Internal helper to detect uniscreen output objects.
}
\keyword{internal}
