% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{perform_categorical_test}
\alias{perform_categorical_test}
\title{Perform categorical test}
\usage{
perform_categorical_test(tab, test_type)
}
\description{
Perform categorical test
}
\keyword{internal}
