% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{perform_continuous_test}
\alias{perform_continuous_test}
\title{Perform continuous tests}
\usage{
perform_continuous_test(var_vec, grp_vec, test_type, stat_type)
}
\description{
Perform continuous tests
}
\keyword{internal}
