% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compfit.R
\name{print.compfit_result}
\alias{print.compfit_result}
\title{Print method showing scoring methodology}
\usage{
\method{print}{compfit_result}(x, ...)
}
\arguments{
\item{x}{Object of class \code{compfit_result}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the input object \code{x}. Called for its
  side effect of printing a formatted summary to the console.
}
\description{
Print method showing scoring methodology
}
\seealso{
Other regression functions:
\code{\link[=compfit]{compfit()}},
\code{\link[=fit]{fit()}},
\code{\link[=fullfit]{fullfit()}},
\code{\link[=multifit]{multifit()}},
\code{\link[=print.fit_result]{print.fit_result()}},
\code{\link[=print.fullfit_result]{print.fullfit_result()}},
\code{\link[=print.multifit_result]{print.multifit_result()}},
\code{\link[=print.uniscreen_result]{print.uniscreen_result()}},
\code{\link[=uniscreen]{uniscreen()}}
}
\concept{regression functions}
\keyword{internal}
