% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifit.R
\name{print.multifit_result}
\alias{print.multifit_result}
\title{Print method for multifit results}
\usage{
\method{print}{multifit_result}(x, ...)
}
\arguments{
\item{x}{Object of class \code{multifit_result}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the input object \code{x}. Called for its
  side effect of printing a formatted summary to the console.
}
\description{
Print method for multifit results
}
\seealso{
Other regression functions:
\code{\link[=compfit]{compfit()}},
\code{\link[=fit]{fit()}},
\code{\link[=fullfit]{fullfit()}},
\code{\link[=multifit]{multifit()}},
\code{\link[=print.compfit_result]{print.compfit_result()}},
\code{\link[=print.fit_result]{print.fit_result()}},
\code{\link[=print.fullfit_result]{print.fullfit_result()}},
\code{\link[=print.uniscreen_result]{print.uniscreen_result()}},
\code{\link[=uniscreen]{uniscreen()}}
}
\concept{regression functions}
\keyword{internal}
