% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtable.R
\name{print.survtable}
\alias{print.survtable}
\title{Print method for survtable}
\usage{
\method{print}{survtable}(x, ...)
}
\arguments{
\item{x}{Object of class \code{survtable}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the input object \code{x}. Called for its
  side effect of printing a formatted summary to the console.
}
\description{
Print method for survtable
}
\seealso{
Other descriptive functions:
\code{\link[=desctable]{desctable()}},
\code{\link[=survtable]{survtable()}}
}
\concept{descriptive functions}
\keyword{internal}
