% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2pptx.R
\name{print.table2pptx_result}
\alias{print.table2pptx_result}
\title{Print method for table2pptx results}
\usage{
\method{print}{table2pptx_result}(x, ...)
}
\description{
Print method for table2pptx results
}
\keyword{internal}
