% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2rtf.R
\name{print.table2rtf_result}
\alias{print.table2rtf_result}
\title{Print method for table2rtf results}
\usage{
\method{print}{table2rtf_result}(x, ...)
}
\description{
Print method for table2rtf results
}
\seealso{
Other export functions: 
\code{\link{autoexport}()},
\code{\link{print.table2docx_result}()},
\code{\link{print.table2pptx_result}()},
\code{\link{table2docx}()},
\code{\link{table2html}()},
\code{\link{table2pdf}()},
\code{\link{table2pptx}()},
\code{\link{table2rtf}()},
\code{\link{table2tex}()}
}
\concept{export functions}
\keyword{internal}
