% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{process_categorical}
\alias{process_categorical}
\title{Process categorical variable}
\usage{
process_categorical(
  data,
  var,
  var_label,
  group_var,
  stats,
  na_include,
  na_label,
  test,
  test_type,
  total,
  total_label,
  na_percent,
  ...
)
}
\description{
Process categorical variable
}
\keyword{internal}
