% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utils.R
\name{validate_predictors_exist}
\alias{validate_predictors_exist}
\title{Validate predictors exist in data}
\usage{
validate_predictors_exist(data, predictors)
}
\arguments{
\item{data}{Data frame}

\item{predictors}{Predictor names}
}
\description{
Validate predictors exist in data
}
\keyword{internal}
