% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clintrial_labels}
\alias{clintrial_labels}
\title{Variable Labels for Clinical Trial Dataset}
\format{
Named character vector with 24 elements
}
\usage{
clintrial_labels
}
\description{
A named character vector providing descriptive labels for all variables
in the clinical_trial dataset. Use with labels parameter in functions.
}
\seealso{
\code{\link{clintrial}}
}
\keyword{datasets}
