% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{format_categorical_stat}
\alias{format_categorical_stat}
\title{Format categorical statistic}
\usage{
format_categorical_stat(n, total, stat_type)
}
\description{
Format categorical statistic
}
\keyword{internal}
