% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{format_continuous_stat}
\alias{format_continuous_stat}
\title{Format continuous statistic}
\usage{
format_continuous_stat(stats, stat_type, fmt_str)
}
\description{
Format continuous statistic
}
\keyword{internal}
