% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{format_pvalues_desctable}
\alias{format_pvalues_desctable}
\title{Format p-values in result table}
\usage{
format_pvalues_desctable(result, p_digits)
}
\description{
Format p-values in result table
}
\keyword{internal}
