% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compfit.R
\name{print.compfit_result}
\alias{print.compfit_result}
\title{Print method showing scoring methodology}
\usage{
\method{print}{compfit_result}(x, ...)
}
\description{
Print method showing scoring methodology
}
\keyword{internal}
