% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifit.R
\name{print.multifit_result}
\alias{print.multifit_result}
\title{Print method for multifit results}
\usage{
\method{print}{multifit_result}(x, ...)
}
\description{
Print method for multifit results
}
\seealso{
Other regression functions: 
\code{\link{compfit}()},
\code{\link{fit}()},
\code{\link{fullfit}()},
\code{\link{multifit}()},
\code{\link{print.compfit_result}()},
\code{\link{print.fit_result}()},
\code{\link{print.fullfit_result}()},
\code{\link{print.uniscreen_result}()},
\code{\link{uniscreen}()}
}
\concept{regression functions}
\keyword{internal}
