% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtable.R
\name{print.survtable}
\alias{print.survtable}
\title{Print method for survtable}
\usage{
\method{print}{survtable}(x, ...)
}
\description{
Print method for survtable
}
\seealso{
Other descriptive functions: 
\code{\link{desctable}()},
\code{\link{survtable}()}
}
\concept{descriptive functions}
\keyword{internal}
