% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2docx.R
\name{print.table2docx_result}
\alias{print.table2docx_result}
\title{Print method for table2docx results}
\usage{
\method{print}{table2docx_result}(x, ...)
}
\description{
Print method for table2docx results
}
\keyword{internal}
