% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2rtf.R
\name{print.table2rtf_result}
\alias{print.table2rtf_result}
\title{Print method for table2rtf results}
\usage{
\method{print}{table2rtf_result}(x, ...)
}
\description{
Print method for table2rtf results
}
\keyword{internal}
