% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniscreen.R
\name{print.uniscreen_result}
\alias{print.uniscreen_result}
\title{Print method for uniscreen results}
\usage{
\method{print}{uniscreen_result}(x, ...)
}
\description{
Print method for uniscreen results
}
\keyword{internal}
