% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtable.R
\name{process_single_outcome}
\alias{process_single_outcome}
\title{Process a single survival outcome}
\usage{
process_single_outcome(
  data,
  outcome,
  outcome_label,
  by,
  times,
  probs,
  stats,
  type,
  conf_level,
  conf_type,
  digits,
  time_digits,
  percent,
  test,
  test_type,
  total,
  total_label,
  time_unit,
  time_label,
  median_label,
  labels,
  na_rm,
  ...
)
}
\description{
Process a single survival outcome
}
\keyword{internal}
