% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_utils.R
\name{reorder_total_column}
\alias{reorder_total_column}
\title{Reorder columns to position total column}
\usage{
reorder_total_column(result, total, total_label)
}
\description{
Reorder columns to position total column
}
\keyword{internal}
