% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utils.R
\name{validate_outcome_exists}
\alias{validate_outcome_exists}
\title{Validate outcome exists in data}
\usage{
validate_outcome_exists(data, outcome)
}
\arguments{
\item{data}{Data frame}

\item{outcome}{Outcome specification}
}
\description{
Validate outcome exists in data
}
\keyword{internal}
